chore(typescript): extracted `primitives/types.d.ts` into individual files
