+++ packages/victory-core/src/victory-primitives/path.d.ts
@@ -1,0 +1,4 @@
+import * as React from "react";
+import { VictoryPrimitiveShapeProps } from "./types";
+
+export default class Path extends React.Component<VictoryPrimitiveShapeProps> {}
